+++ src/main/java/com/gonggoo/gonggoo/auth/oauth/google/GoogleConfig.java
@@ -1,4 +1,4 @@
-package com.gonggoo.gonggoo.auth.google;
+package com.gonggoo.gonggoo.auth.oauth.google;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
