+++ src/main/java/com/gonggoo/gonggoo/auth/oauth/google/GoogleProps.java
@@ -1,13 +1,14 @@
-package com.gonggoo.gonggoo.auth.kakao;
+package com.gonggoo.gonggoo.auth.oauth.google;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-@ConfigurationProperties(prefix = "oauth.kakao")
-public record KakaoProps(
+@ConfigurationProperties(prefix = "oauth.google")
+public record GoogleProps(
         String client_id,
         String redirect_uri,
         String client_secret,
         String post_uri,
-        String user_info_url
+        String user_info_url,
+        String scope
 ) {
 }
