+++ src/main/java/com/gonggoo/gonggoo/auth/oauth/kakao/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package com.gonggoo.gonggoo.auth.kakao;
+package com.gonggoo.gonggoo.auth.oauth.kakao;
 
 public record KakaoTokenResponse(
         String token_type,
