+++ src/main/java/com/gonggoo/gonggoo/auth/oauth/naver/NaverProps.java
@@ -1,14 +1,13 @@
-package com.gonggoo.gonggoo.auth.google;
+package com.gonggoo.gonggoo.auth.oauth.naver;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-@ConfigurationProperties(prefix = "oauth.google")
-public record GoogleProps(
+@ConfigurationProperties(prefix = "oauth.naver")
+public record NaverProps(
         String client_id,
         String redirect_uri,
         String client_secret,
         String post_uri,
-        String user_info_url,
-        String scope
+        String user_info_url
 ) {
 }
