+++ src/main/java/com/gonggoo/gonggoo/auth/oauth/naver/NaverTokenResponse.java
@@ -1,4 +1,4 @@
-package com.gonggoo.gonggoo.auth.naver;
+package com.gonggoo.gonggoo.auth.oauth.naver;
 
 public record NaverTokenResponse(
         String access_token,
