+++ src/main/java/com/gonggoo/gonggoo/auth/service/AuthService.java
@@ -10,17 +10,17 @@ import static com.gonggoo.gonggoo.global.response.ErrorCode.PASSWORD_FAILED;
 import com.gonggoo.gonggoo.auth.domain.RegistrationProvider;
 import com.gonggoo.gonggoo.auth.dto.request.LoginRequest;
 import com.gonggoo.gonggoo.auth.dto.response.LoginResponse;
-import com.gonggoo.gonggoo.auth.google.GoogleProps;
-import com.gonggoo.gonggoo.auth.google.GoogleTokenResponse;
-import com.gonggoo.gonggoo.auth.google.GoogleUserInfoResponse;
+import com.gonggoo.gonggoo.auth.oauth.google.GoogleProps;
+import com.gonggoo.gonggoo.auth.oauth.google.GoogleTokenResponse;
+import com.gonggoo.gonggoo.auth.oauth.google.GoogleUserInfoResponse;
 import com.gonggoo.gonggoo.auth.jwt.JwtTokenDto;
 import com.gonggoo.gonggoo.auth.jwt.JwtTokenProvider;
-import com.gonggoo.gonggoo.auth.kakao.KakaoProps;
-import com.gonggoo.gonggoo.auth.kakao.KakaoTokenResponse;
-import com.gonggoo.gonggoo.auth.kakao.KakaoUserInfoResponse;
-import com.gonggoo.gonggoo.auth.naver.NaverProps;
-import com.gonggoo.gonggoo.auth.naver.NaverTokenResponse;
-import com.gonggoo.gonggoo.auth.naver.NaverUserInfoResponse;
+import com.gonggoo.gonggoo.auth.oauth.kakao.KakaoProps;
+import com.gonggoo.gonggoo.auth.oauth.kakao.KakaoTokenResponse;
+import com.gonggoo.gonggoo.auth.oauth.kakao.KakaoUserInfoResponse;
+import com.gonggoo.gonggoo.auth.oauth.naver.NaverProps;
+import com.gonggoo.gonggoo.auth.oauth.naver.NaverTokenResponse;
+import com.gonggoo.gonggoo.auth.oauth.naver.NaverUserInfoResponse;
 import com.gonggoo.gonggoo.global.domain.Role;
 import com.gonggoo.gonggoo.global.exception.NeighborsException;
 import com.gonggoo.gonggoo.member.domain.Member;
